docs: Add Storages & Request loaders pages
