+++ docs/guides/code/storages/kvs_basic_example.py
