+++ docs/guides/code/storages/rq_with_crawler_example.py
