+++ website/package.json
@@ -40,6 +40,7 @@
         "@docusaurus/mdx-loader": "^3.5.2",
         "@docusaurus/plugin-client-redirects": "^3.5.2",
         "@docusaurus/preset-classic": "^3.5.2",
+        "@docusaurus/theme-mermaid": "^3.5.2",
         "@giscus/react": "^3.0.0",
         "@mdx-js/react": "^3.0.1",
         "axios": "^1.5.0",
